pldmtool:pdr fix type cast and data usage
  - Cast transferFlag to uint8_t for consistent enum comparison
  - Use recordData.size() instead of respCnt for accurate length
  - Replace respRecordData with recordData.data() for placement new
  - Ensure correct parsing of pldm_pdr_hdr from received data

Tested:
'''
{
    "nextRecordHandle": 0,
    "responseCount": 116,
    "recordHandle": 1,
    "PDRHeaderVersion": 1,
    "PDRType": "Redfish Resource PDR",
    "recordChangeNumber": 0,
    "dataLength": 106,
    "ResourceID": 1,
    "ResourceFlags": 1,
    "ContainingResourceID": 0,
    "ProposedContainingResourceLengthBytes": 22,
    "ProposedContainingResourceName": "Processors.Processors",
    "SubURILengthBytes": 31,
    "SubURI": "Oem/AMD/SocConfiguration/Token",
    "AdditionalResourceIDCount": 0,
    "MajorSchemaVersion": "00.00.00",
    "MajorSchemaDictionaryLengthBytes": 481,
    "MajorSchemaDictionarySignature": 2300451081,
    "MajorSchemaNameLength": 25,
    "MajorSchemaName": "AmdSocConfigurationToken",
    "OEMCount": 0
}
'''

+++ pldmtool/pldm_platform_cmd.cpp
@@ -604,8 +604,11 @@ class GetPDR : public CommandInterface
                               respRecordData + respCnt);
 
             // End or StartAndEnd
-            if (transferFlag == PLDM_PLATFORM_TRANSFER_END ||
-                transferFlag == PLDM_PLATFORM_TRANSFER_START_AND_END)
+            if ((transferFlag ==
+                     static_cast<uint8_t>(PLDM_PLATFORM_TRANSFER_END) ||
+                 transferFlag == static_cast<uint8_t>(
+                                     PLDM_PLATFORM_TRANSFER_START_AND_END)))
+
             {
                 printPDRMsg(nextRecordHndl, recordData.size(),
                             recordData.data(), terminusHandle);
